pyproject.toml
[project]
name = "stock-sentiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.6",
    "matplotlib>=3.10.5",
    "newsapi-python>=0.2.7",
    "pyfiglet>=1.0.3",
    "rich>=14.1.0",
    "scipy>=1.16.1",
    "yfinance>=0.2.65",
]
